Add files via upload
Based on ML algorithm for the analysis and detection of potential fraud in a bank where customers are applying for the advance credit cards. This dataset is based on that data in which the customer has to provide when filling the application forum.My objective is to detect potential fraud within the applications.
This project is based on Unsupervised Deep Learning in which it is going to identify some patterns in a high dimensional dataset with full non linearity relationships. I am using here Self-Organizing-Maps.